+++ packages/fastboard-react/src/index.tsx
@@ -13,7 +13,7 @@ export * from "./ZoomControl";
 export * from "./PageControl";
 export * from "./Toolbar";
 export * from "./Fastboard";
-export { useFastboard } from "./hooks";
+export { useFastboard, useReplayFastboard } from "./hooks";
 
 import "./style.scss";